+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature #291: add support for sending 'raw' requests (thanks to jachstet-sea and chuchuva)
 - Feature #155: add OPTIONS method
+- Feature #283: improve error message on timeout on browser platform
 
 ## 2.3.1